TASK-1325177: Certificates to leverage ESO (#448)
TASK-1325177: Update truststore certificate to leverage ESO
