--- charts/pega/templates/pega-certificates-config.yaml
@@ -1,1 +1,0 @@
-{{- include "pegaCertificatesConfigTemplate" . }}
