+++ charts/pega/templates/pega-certificates-secret.yaml
@@ -1,0 +1,1 @@
+{{- include "pegaCertificatesSecretTemplate" . }}
